requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
appdirs==1.4.4
async-timeout==4.0.1
attrs @ file:///tmp/build/80754af9/attrs_1620827162558/work
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
black==19.10b0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
configparser==5.2.0
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
debugpy @ file:///home/conda/feedstock_root/build_artifacts/debugpy_1636043249011/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1631346842025/work
diffusion-utils @ file:///home/jack/diffusion-utils
docker-pycreds==0.4.0
docstring-parser==0.13
einops==0.3.2
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1605121927639/work/dist/entrypoints-0.3-py2.py3-none-any.whl
fonttools==4.25.0
frozenlist==1.2.0
fsspec==2021.11.1
future==0.18.2
gitdb==4.0.9
GitPython==3.1.24
google-auth==2.3.3
google-auth-oauthlib==0.4.6
grpcio==1.42.0
idna==3.3
importlib-metadata==4.8.3
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1638555504864/work/dist/ipykernel-6.6.0-py3-none-any.whl
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1638470227058/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1637175083648/work
jsonargparse==4.1.0
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1637611911738/work
jupyter-core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1636814260563/work
kiwisolver @ file:///tmp/build/80754af9/kiwisolver_1612281846643/work
kornia==0.6.2
lpips==0.1.4
Markdown==3.3.6
matplotlib @ file:///tmp/build/80754af9/matplotlib-suite_1634667019719/work
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1631080358261/work
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
multidict==5.2.0
munkres==1.1.4
mypy-extensions==0.4.3
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1638419302549/work
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1634095651905/work
oauthlib==3.1.1
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
packaging==21.3
pandas==1.3.5
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pathspec==0.7.0
pathtools==0.1.2
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1602535608087/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow @ file:///home/conda/feedstock_root/build_artifacts/pillow_1636558800840/work
promise==2.3
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1639065841292/work
protobuf==3.19.1
psutil==5.8.0
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.1
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1629119114968/work
pyparsing==3.0.6
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytorch-lightning==1.5.6
pytz==2021.3
PyYAML==6.0
pyzmq @ file:///home/conda/feedstock_root/build_artifacts/pyzmq_1635877397296/work
regex @ file:///tmp/build/80754af9/regex_1628063347754/work
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
scipy==1.7.3
sentry-sdk==1.5.1
shortuuid==1.0.8
sip==4.19.13
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smmap==5.0.0
subprocess32==3.5.4
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
torch==1.10.1
torch-tb-profiler==0.3.1
torchmetrics==0.6.2
torchvision==0.11.2
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1635819584296/work
tqdm==4.62.3
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1635260543454/work
typed-ast @ file:///tmp/build/80754af9/typed-ast_1624953673314/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1638334978229/work
urllib3==1.26.7
vector-quantize-pytorch @ git+https://github.com/GallagherCommaJack/vector-quantize-pytorch@7de68879804fdc6a747c1cc9bae881a4e8a3995b
wandb==0.12.8
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1600965781394/work
Werkzeug==2.0.2
yarl==1.7.2
yaspin==2.1.0
zipp==3.6.0
